pyproject.toml
[project]
name = "synapse-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Tharun Ganapathi",email = "tharunganapathi19@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "youtube-transcript-api (>=1.2.3,<2.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "genanki (>=0.13.1,<0.14.0)",
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pypdf (>=6.5.0,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pdfplumber (>=0.11.8,<0.12.0)",
    "google-generativeai (>=0.8.6,<0.9.0)"
]

[tool.poetry]
packages = [{include = "synapse"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
